requirements.txt
Flask==2.3.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
PyYAML==6.0
boto3==1.28.25
mlflow==2.5.0
optuna==3.3.0
